more details for clarity, changed to mock decorators, small fixes
