+++ amz_review_scraper/models.py
@@ -1,6 +1,6 @@
 from app import db
 
-# Maybe change this to Item?
+
 class Item(db.Model):
     __tablename__ = "items"
     id = db.Column(db.Integer, primary_key=True)
